coap: Parse packages without a payload marker
The payload marker is only mandatory in a package if there is actual
payload. The option iteration therefore can't rely on its presence but
has to keep track of the end-of-message pointer -- that's a word more to
carry around, but we can't do without it unless we'd deliberately place
a payload marker in the message, and with data allocaetd by a network
stack, we generally can't do that.

As a side effect, unsafe code is reduced.